+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 - linking tree view and map hover
 - auto fit scene button
 - anugularJS material
+- Added Link in visualization #84
 
 ### Changed